export getHandler instead of wrappers
