+++ lib/rest/getHandler.js
@@ -39,7 +39,7 @@ function makeArgsGetter (argsOrder) {
 
 // A function that returns the middleware for a given method and service
 // `getArgs` is a function that should return additional leading service arguments
-function getHandler (method) {
+module.exports = function getHandler (method) {
   return service => {
     const getArgs = makeArgsGetter(service.methods[method]);
 
@@ -99,13 +99,4 @@ function getHandler (method) {
         });
     };
   };
-}
-
-module.exports = {
-  find: getHandler('find'),
-  get: getHandler('get'),
-  create: getHandler('create'),
-  update: getHandler('update'),
-  patch: getHandler('patch'),
-  remove: getHandler('remove')
 };
