+++ package.json
@@ -48,14 +48,13 @@
     "lib": "lib"
   },
   "dependencies": {
-    "@feathersjs/commons": "^1.4.0",
+    "@feathersjs/commons": "^2.0.0",
     "@feathersjs/errors": "^3.0.0",
     "debug": "^3.1.0",
     "express": "^4.16.2",
     "uberproto": "^2.0.0"
   },
   "devDependencies": {
-    "@feathersjs/feathers": "^3.0.5",
     "axios": "^0.18.0",
     "istanbul": "^1.1.0-alpha.1",
     "mocha": "^5.0.0",
